backend/requirements.txt
# FastAPI и зависимости
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# WebSocket
websockets==12.0

# Pydantic для валидации
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Дополнительно
httpx==0.26.0
redis==5.0.1
aiosmtplib==3.0.1
beautifulsoup4==4.12.3
lxml==5.1.0
playwright==1.49.1
playwright-stealth==1.0.6
